==> 01_Sat_English_Solve/code_ipynb/01_data_preprocess.ipynb <==
!pip install ipynb-py-convert

!ipynb-py-convert 05_sat_demo.ipynb 05_sat_demo.py

!jupyter nbconvert --to script 01_data_preprocess.ipynb

import warnings
warnings.simplefilter(action='ignore')

import pandas as pd

## SAT datset

sat_data = pd.read_csv("data/raw/sat_problems.csv")
sat_data = sat_data[sat_data["question"].map(lambda x: "어법" in x)]

sat_train = sat_data[(sat_data["year"] < 2020)]
sat_valid = sat_data[(sat_data["year"] == 2020)]
sat_test = sat_data[sat_data["year"] > 2020]

print("# of train data:", sat_train.shape[0])
print("# of valid data:", sat_valid.shape[0])
print("# of test data:", sat_test.shape[0])

def clean_bracket(string):
    string = string.replace("[", "")
    string = string.replace("]", "")
    return string

sat_train["context"] = sat_train["context"].map(clean_bracket)
sat_valid["context"] = sat_valid["context"].map(clean_bracket)
sat_test["context"] = sat_test["context"].map(clean_bracket)

sat_train["label"] = sat_train["label"].map(int)
sat_valid["label"] = sat_valid["label"].map(int)
sat_test["label"] = sat_test["label"].map(int)

sat_train.to_csv("data/processed/sat_train.tsv", sep="\t", index=False)
sat_valid.to_csv("data/processed/sat_valid.tsv", sep="\t", index=False)
sat_test.to_csv("data/processed/sat_test.tsv", sep="\t", index=False)

## CoLA Daaset

columns = ["source", "label", "original_judgement", "context"]
in_domian_train = pd.read_csv("data/raw/cola/in_domain_train.tsv", sep="\t", header=None, names=columns)
in_domian_dev = pd.read_csv("data/raw/cola/in_domain_dev.tsv", sep="\t", header=None, names=columns)
out_of_domian_dev = pd.read_csv("data/raw/cola/out_of_domain_dev.tsv", sep="\t", header=None, names=columns)

print("# of cola train data:", in_domian_train.shape[0])
print("# of cola valid data:", in_domian_dev.shape[0])
print("# of cola test data:", out_of_domian_dev.shape[0])

in_domian_train[["context", "label"]].to_csv("data/processed/cola_train.tsv", sep="\t", index=False)
in_domian_dev[["context", "label"]].to_csv("data/processed/cola_valid.tsv", sep="\t", index=False)
out_of_domian_dev[["context", "label"]].to_csv("data/processed/cola_test.tsv", sep="\t", index=False)
